+++ library/irods_rulesets.py
@@ -42,10 +42,12 @@ def main():
             server_config['re_rulebase_set'] = []
             for ruleset in rulesets:
                 server_config['re_rulebase_set'].append({'filename': ruleset})
-            data_file.seek(0)
-            json.dump(server_config, data_file, indent=4, sort_keys=True)
-            data_file.truncate()
             changed = True
+
+            if not module.check_mode:
+                data_file.seek(0)
+                json.dump(server_config, data_file, indent=4, sort_keys=True)
+                data_file.truncate()
 
     module.exit_json(
             changed=changed,
